feat: remove reliance on revised library using getControl0()
I had created a new function called getControl0() in the
ES100_Arduino_library, so I could dump out the value of
the ES100 Control0 register when I thought something was
not quite right.  Everything was fine, but that means you
must use that version of ES100_Arduino_library with my
code.  By eliminating the reference to getControl0(), the
code is now compatible with the original library.

Signed-off-by: Dean Weiten <dmw@weiten.com>